chore(rtdb): Exposed admin.database namespace
