+++ src/firebase-namespace.d.ts
@@ -16,5 +16,6 @@
 
 export * from './credential/index';
 export * from './firebase-namespace-api';
+export * from './database/index';
 export * from './messaging/index';
 export * from './remote-config/index';
